requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.23.0
jinja2>=3.1.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
sqlalchemy>=2.0.0
anthropic>=0.25.0
